add policy recommend and refactor policy Constant
